requirements.txt
click==8.0.3
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.21.3
Pillow==8.4.0
